Fix crash when returning a non string value
Fix a crash caused by a strict return type (string) which can occurs when an argument is an integer.
This argument can be an integer if the product name is composed of number only. It is casted to an integer because it is used as an array key when building the format's arguments.